Some old API breaking changes
* bot.players is now a lookup function, the caching of entities was bad.
* entity's now have an isValid property, plugins should check this.
Ignore an entity if it's not valid (Set to false when destroyed)
* bot.players takes a string, regular expression or function to filter
your lookups, quite handy. No argument will return all players. Most
arguments will return an Array of entities, except for a String when it
matches exactly to a player's name, in which case just that player is
returned.
* Player Entities are now not a thing, they are just entities, player
information is stored in the entity property, entity.playerInfo. EG.
entity.playerInfo.ping