Clarified when to use login and spawn
It said spawn was fired when you load in for the first, but this is false. And at login it refered to spawn to use when you want to do something as soon as you  join the server, but as spawn isn't fired when you join the server this was false too.