Fix bot.isSleeping being desync
As the packet 'bed' has been removed since 1.14, bot.isSleeping was desynced, which caused bots to not be able to sleep when they could have and/or to not be able to wake when they could have.
This has been done by listening for the "pose" property that is emitted in entity_metadata.