Promisify testCommon   (#1586)
* Promisify testCommon  I have chosen not to promisify all the tests themselves for now. All bot.test methods use callbackify such that they are still compatible with every existing test, but they all return promises now so that they may be used with future tests as well.

* Add better test resolution in externalTest

Previously, tests would be resolved based on the current working directory (where you started the tests from), rather than where the externalTest.js file is located on the file system. The timeout was also not being set for individual tests.

* Add new promise utils

withTimeout attaches a timeout to a promise and causes it to reject after a certain amount of time. onceWithCleanup is similar to once, but does listener removal and also supports timeouts/conditionals.

* convert all "new Promise"s to use onceWithCleanup

* Fix call stack in onceWithCleanup

* Always resolve creative mode promise

* Use an async function for setCreativeMode

.then(resolve, resolve) rubbed me the wrong way

* withTimeout now uses Promise.race, add typeof checks

* Bump test timeout to 60 seconds

Fishing takes a -really- long time (around 45 seconds on average in my testing), so 30s doesn't cut it.