Optimize findblock distance filter
by not using `distanceTo()`, we dont use `Math.sqrt()` which is costly in processing. instead we square the maxDistance once before anything else and use distance squared. This works because we are only comparing distances, not performing arithmetic or using it for anything significant.

This also works when sorting the blocks by distance because we only need a positive value, negative value, or 0. Getting the difference between their exact euclidian distances is overkill.