Bypass anticheats that detect sensitivity
Anticheats can detect the player's sensitivity by comparing how much the player's look direction changes each tick and getting the GCD of that. In vanilla when using 100% sensitivity, the player always moves by a multiple .15 when they're manually moving the camera with the mouse. This change makes mineflayer do that too. My own proof-of-concept of an anticheat that does this is on my very unfinished server (matdoes.dev) but there's likely other anticheats that do this as well.